[BE] 스프링 서버 배포 전 BE/develop 진행 상황 반영

Closes S09P12C202-400, S09P12C202-352, and S09P12C202-333

See merge request s09-webmobile1-sub2/S09P12C202!122

+++ backend/planty-service/src/main/java/com/planty/db/entity/GmInfo.java
@@ -21,7 +21,7 @@ public class GmInfo {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "gid", nullable = false) // 식별키
-    private Integer gid;
+    private Long gid;
 
     @NonNull
     @Column(name ="id", length = 16, nullable = false) // 아이디
